Query edition wizard
It is possible to edit an existing query from the task list by
selecting a query and then after a right-click selecting the
Properties menu.
If the query is a default query, the default queries wizard is
opened with the list of predefined reports.
If the query is a custom query, the custom page is opened with all
the fields defined by the configuration.