Add the first version of the server REST connection.
A first version of the connection with the server has been contributed.
In this version, we are depending on Restlet to communicate with the
server.

The mock server has been updated to fix an issue with HTTP headers and
case sensitivity (the issue is not completely fixed everywhere) and a
loag has been added to have a better view of the requests from the
clients.