Add support for a debug mode in Tuleap.
A new preference page has been introduced in order to let end users
activate a debug mode in which the Mylyn Connector for Tuleap will log
additional information. This debug mode is used by the connector in
order to log the content of the TaskData when a Tuleap task is being
opened.

Change-Id: I468c59ca3dc4014af22c0c3367c421dda58400cd